session delete

+++ Controllers/Login.php
@@ -36,10 +36,12 @@ class Login
     {
       $cur_session_id = \Session::getId();
       \m_SessionShare::where('session_id',$cur_session_id)->delete();
-      \Session::forget('session_share_set');
-      
-      \DB::connection('service_users')->table('sessions')
-        ->where('id',$cur_session_id)->delete();
+      if (\SettingsSite::get('system_use_sso') == 'on') {
+        \Session::forget('session_share_set');
+        
+        \DB::connection('service_users')->table('sessions')
+          ->where('id',$cur_session_id)->delete();
+      }
 
       Auth::logout();
       return redirect(request()->getSchemeAndHttpHost().'/');
